Handle revealed state with simply removing Doors from the list
This is closer to the original solution, and makes sense. Since we don't
actually care what index a door was before we remove it. And maintaining
it in the list simply makes choosing the next door less efficient.

The selection of the hosts door to reveal is also simplified, instead of choosing it by random, we simply remove the first Goat door. Perhaps that impacts the accuracy of this a little. But ultimately the contestant would still be presented with the same probability in the second choice.